% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_content.R
\name{information_content}
\alias{information_content}
\title{Compute the Information Content}
\usage{
information_content(peps)
}
\arguments{
\item{peps}{A character vector of equal length peptides}
}
\value{
A list with 3 elements: i. \code{IC_pos}: Named numeric vector with
the computed positional Information Content, ii. \code{IC_mat}: A
numeric matrix with computed Information Content contribution from
each residue at each position, iii. \code{IC_logo_aa_height}: The computed
Information Content for logo visualisation, such that
\code{X[i,j] = frequency_matrix[i,j] * IC_pos[i] * sign(IC_mat[i,j]}
}
\description{
Compute the Information Content
}
\examples{
information_content(peps = PEPTIDES)

}
