% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kullback_leibler.R
\name{kullback_leibler}
\alias{kullback_leibler}
\title{Compute the Kullback-Leibler Divergence}
\usage{
kullback_leibler(peps, aa_background_distribution = "All")
}
\arguments{
\item{peps}{A character vector of equal length peptides}

\item{aa_background_distribution}{One of 'Viruses', 'Archaea', 'Bacteria',
'Eukaryota', 'All', where 'ALL' is default. See ?BGFREQS for more.}
}
\value{
A list with 3 elements: i. \code{KLD_pos}: Named numeric vector with
the computed positional Kullback-Leibler Divergence, ii. \code{KLD_mat}: A
numeric matrix with computed Kullback-Leibler Divergence contribution from
each residue at each position, iii. \code{KLD_logo_aa_height}: The computed
Kullback-Leibler Divergence for logo visualisation, such that
\code{X[i,j] = frequency_matrix[i,j] * KLD_pos[i] * sign(KLD_mat[i,j]}
}
\description{
Compute the Kullback-Leibler Divergence
}
\examples{
kullback_leibler(peps = PEPTIDES, aa_background_distribution = 'All')

}
