% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_chop.R
\name{pep_chop}
\alias{pep_chop}
\title{Chop a vector of peptides into k-mers}
\usage{
pep_chop(x, k)
}
\arguments{
\item{x}{A vector of peptides}

\item{k}{length of k-mers}
}
\value{
A list of the possible k-mers generated for each peptide
}
\description{
Chop a vector of peptides into k-mers
}
\examples{

peps = pep_ran(k = 9, n = 10)
kmers = pep_chop(x = peps, k = 3)

}
