% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_extend.R
\name{pep_extend}
\alias{pep_extend}
\title{Extend peptides in a vector of peptides}
\usage{
pep_extend(pep, k = NULL, ext_char = "X")
}
\arguments{
\item{x}{A character vector of peptides to be extended}
}
\value{
A character vector of peptides of equal length
}
\description{
Given a character vector of peptides of unequal length, pep_extend will add
'X to the end (C-terminus) of the peptides, such that any peptide shorter
than the longest peptide will be extended to match the length
}
\examples{
peps = c('NWAWK','TATIDISM','KGLFEPWRCRC','MNSMEEMRESG',
         'HPQRHCHAQWYEEG','KIHSELVSVSVWA','WAGDIHHVC','NHSTGPKEGGFIMY',
         'GGFDVHIRWHTYKCP','ITHSPNMLSMVDT')
nchar(peps)
peps_ext = pep_extend(peps)
nchar(peps_ext)

}
