% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_mutate.R
\name{pep_mutate}
\alias{pep_mutate}
\title{Mutate residues along peptide positions}
\usage{
pep_mutate(pep, pos, res, keep_wt = FALSE)
}
\arguments{
\item{pep}{A string representing a peptide}

\item{pos}{Which position(s) to mutate}

\item{res}{Which amino acid(s) to mutate to}

\item{keep_wildtype}{Boolean should \code{pep} be included in the output}
}
\value{
A character vector with number of elements corresponding to the
number of peptides required to cover all mutations
}
\description{
This is equivalent to making an alanine scan, where a given peptide is
substituted positionally with 'A'. Note, unique peptides are returned.
}
\examples{
pep_mutate(pep = "RQGQDHPTM", pos = seq(1,9), res = AMINOACIDS$one, keep_wt = FALSE)
length(pep_mutate(pep = "RQGQDHPTM", pos = seq(1,9), res = AMINOACIDS$one, keep_wt = FALSE))
pep_mutate(pep = "RQGQDHPTM", pos = seq(1,9), res = AMINOACIDS$one, keep_wt = TRUE)
length(pep_mutate(pep = "RQGQDHPTM", pos = seq(1,9), res = AMINOACIDS$one, keep_wt = TRUE))

}
