% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_scale.R
\name{range_scale}
\alias{range_scale}
\title{Linearly scales values to new range}
\usage{
range_scale(x, new_min = -1, new_max = 1)
}
\arguments{
\item{x}{A set of numbers as a numeric vector, matrix or data frame}
}
\value{
The scaled values in same structure as was input
}
\description{
Linearly scales values to new range
}
\examples{
sample = rnorm(10)
range_scale(sample)

}
