% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Compute the Shannon Entropy}
\usage{
shannon_entropy(peps)
}
\arguments{
\item{peps}{A character vector of equal length peptides}
}
\value{
A list with 3 elements: i. \code{H_pos}: Named numeric vector with
the computed positional Shannon Entropy, ii. \code{H_mat}: A
numeric matrix with computed Shannon Entropy contribution from
each residue at each position, iii. \code{H_logo_aa_height}: The computed
Shannon Entropy for logo visualisation, such that
\code{X[i,j] = frequency_matrix[i,j] * H_pos[i] * sign(H_mat[i,j]}
}
\description{
Compute the Shannon Entropy
}
\examples{
shannon_entropy(peps = PEPTIDES)

}
